requirements.txt
fastapi>=0.115.0
uvicorn>=0.20.0
pydantic>=2.0.0
python-dotenv>=1.0.0
qdrant-client>=1.10.0
openai>=1.50.0
tiktoken>=0.8.0
python-multipart>=0.0.20